+++ test/fixtures/jsdocAncestorSomeTag.js
@@ -294,7 +294,7 @@ const body = {
         typeLines: []
       }
     ],
-    initial: '',
+    initial: '      ',
     type: 'JsdocBlock'
   },
   loc: {