+++ test/fixtures/jsdocAncestorSomeTagTS.js
@@ -283,7 +283,7 @@ const body = {
         typeLines: []
       }
     ],
-    initial: '',
+    initial: '      ',
     type: 'JsdocBlock'
   },
   loc: {