+++ test/fixtures/jsdocSomeTag.js
@@ -106,7 +106,7 @@ const body = {
         typeLines: []
       }
     ],
-    initial: '',
+    initial: '      ',
     type: 'JsdocBlock'
   },
   loc: {